+++ packages/bsky/src/api/app/bsky/unspecced/getSuggestedStarterPacks.ts
@@ -27,7 +27,7 @@ export default function (server: Server, ctx: AppContext) {
   )
   server.add(app.bsky.unspecced.getSuggestedStarterPacks, {
     auth: ctx.authVerifier.standardOptional,
-    handler: async ({ auth, params, req }) => {
+    handler: async ({ auth, params, req, signal }) => {
       const viewer = auth.credentials.iss
       const labelers = ctx.reqLabelers(req)
       const hydrateCtx = await ctx.hydrator.createContext({ labelers, viewer })
@@ -42,6 +42,7 @@ export default function (server: Server, ctx: AppContext) {
           ...params,
           hydrateCtx,
           headers,
+          signal,
         },
         ctx,
       )
@@ -71,6 +72,7 @@ const skeleton = async (
     },
     {
       headers: params.headers,
+      signal: params.signal,
     },
   )
 
@@ -140,6 +142,7 @@ type Context = {
 type Params = app.bsky.unspecced.getSuggestedStarterPacks.$Params & {
   hydrateCtx: HydrateCtx & { viewer: string | null }
   headers: Record<string, string>
+  signal: AbortSignal
 }
 
 type SkeletonState = {