+++ packages/ozone/src/context.ts
@@ -34,6 +34,7 @@ import {
   ReportStatsService,
   type ReportStatsServiceCreator,
 } from './report/stats.js'
+import { SafeDidResolver } from './safe-fetch.js'
 import {
   SafelinkRuleService,
   type SafelinkRuleServiceCreator,
@@ -131,6 +132,12 @@ export class AppContext {
       plcUrl: cfg.identity.plcUrl,
       didCache,
     })
+    if (!cfg.service.devMode) {
+      idResolver.did = new SafeDidResolver({
+        plcUrl: cfg.identity.plcUrl,
+        didCache,
+      })
+    }
 
     const createAuthHeaders = (aud: string, lxm: string) =>
       createServiceAuthHeaders({
@@ -145,6 +152,7 @@ export class AppContext {
       ? new BlobDiverter(db, {
           idResolver,
           serviceConfig: cfg.blobDivert,
+          devMode: cfg.service.devMode,
         })
       : undefined
     const eventPusher = new EventPusher(db, createAuthHeaders, {