+++ packages/lex/lex-resolver/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @atproto/lex-resolver
 
+## 0.2.10
+
+### Patch Changes
+
+- Updated dependencies [[`ec29fcd`](https://github.com/bluesky-social/atproto/commit/ec29fcdb4e0ce098fb1473feebe876a94b1c8e33)]:
+  - @atproto/lex-client@0.3.5
+
 ## 0.2.9
 
 ### Patch Changes