+++ packages/ozone/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @atproto/ozone
 
+## 0.3.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @atproto/lex@0.3.8
+  - @atproto/xrpc-server@0.12.7
+
 ## 0.3.0
 
 ### Minor Changes